modifié :         .vscode/settings.json         modifié :         Makefile         modifié :         boards/platforms/digilent_arty_poto.py         modifié :         boards/targets/digilent_arty_cv32e41p.py         nouveau fichier : examples/ReadMe_zephyr         nouveau fichier : examples/demo-lora/.vscode/settings.json         modifié :         examples/demo-lora/Makefile         nouveau fichier : examples/demo-lora/all_test.c         modifié :         examples/demo-lora/demo.elf.map         modifié :         examples/demo-lora/lora/boards/board.c         modifié :         examples/demo-lora/lora/boards/rtc-board.c         modifié :         examples/demo-lora/lora/boards/spi-board.c         modifié :         examples/demo-lora/lora/system/spi.h         modifié :         examples/demo-lora/main.c         modifié :         examples/demo-lora/test1.c         modifié :         examples/demo-lora/testspi.c         nouveau fichier : examples/demo-lora/tools/archive_test/libspi.c         nouveau fichier : examples/demo-lora/tools/archive_test/libspi.h         modifié :         examples/demo-lora/tools/console.c         modifié :         examples/demo-lora/tools/console.h         modifié :         examples/demo-lora/tools/isr.c         modifié :         examples/demo-lora/tools/isr.h         modifié :         examples/demo-lora/tools/libdio.c         modifié :         examples/demo-lora/tools/libdio.h         modifié :         examples/demo-lora/tools/libtimer.c         modifié :         examples/demo-lora/tools/libtimer.h         modifié :         examples/demo/Makefile         nouveau fichier : examples/demo/all_test.c         modifié :         examples/demo/demo.elf.map         modifié :         examples/demo/iotest.c         modifié :         examples/demo/lora/boards/board-config.h         modifié :         examples/demo/lora/boards/board.c         modifié :         examples/demo/lora/boards/gpio-board.c         modifié :         examples/demo/lora/boards/i2c-board.c         modifié :         examples/demo/lora/boards/rtc-board.c         modifié :         examples/demo/lora/boards/rtc-board.h         modifié :         examples/demo/lora/boards/spi-board.c         modifié :         examples/demo/lora/boards/sx1276-board.c         modifié :         examples/demo/lora/boards/sx1276-board.h         supprimé :        examples/demo/lora/boards/sx1276mb1las-board.c         supprimé :        examples/demo/lora/boards/sx1276mb1mas-board.c         modifié :         examples/demo/lora/boards/uart-board.c         modifié :         examples/demo/lora/radio/sx1276.c         modifié :         examples/demo/lora/radio/sx1276.h         modifié :         examples/demo/main.c         nouveau fichier : examples/demo/ping_pong/ping_pong.ino         modifié :         examples/demo/simple.c         modifié :         examples/demo/test1.c         modifié :         examples/demo/testspi.c         nouveau fichier : examples/demo/tools/archive_test/Makefile         nouveau fichier : examples/demo/tools/archive_test/librtc_hw.c         nouveau fichier : examples/demo/tools/archive_test/librtc_hw.h         nouveau fichier : examples/demo/tools/archive_test/librtc_stm.c         nouveau fichier : examples/demo/tools/archive_test/librtc_stm.h         modifié :         examples/demo/tools/binary.h         modifié :         examples/demo/tools/console.c         modifié :         examples/demo/tools/console.h         modifié :         examples/demo/tools/defint.h         modifié :         examples/demo/tools/delai.c         modifié :         examples/demo/tools/delai.h         modifié :         examples/demo/tools/isr.c         modifié :         examples/demo/tools/isr.h         nouveau fichier : examples/demo/tools/ld/linker.ld         nouveau fichier : examples/demo/tools/libdio.c         nouveau fichier : examples/demo/tools/libdio.h         supprimé :        examples/demo/tools/libisr.backupp         supprimé :        examples/demo/tools/libisr.c         supprimé :        examples/demo/tools/libisr.h         modifié :         examples/demo/tools/libspi.c         modifié :         examples/demo/tools/libspi.h         nouveau fichier : examples/demo/tools/libtimer.c         nouveau fichier : examples/demo/tools/libtimer.h         nouveau fichier : examples/demo/tools/test_fonction/deadTimeAlarmTEST.c         nouveau fichier : examples/demo/tools/test_fonction/timeStampTEST.c         nouveau fichier : examples/generated.sh         nouveau fichier : examples/test.sh         nouveau fichier : examples/zephyr-on-litex-vexriscv/install.sh         nouveau fichier : examples/zephyr-on-litex-vexriscv/install_litex.sh         nouveau fichier : examples/zephyr-on-litex-vexriscv/install_toolchain.sh

+++ examples/demo-lora/lora/boards/spi-board.c
@@ -45,7 +45,7 @@ void SpiDeInit(Spi_t *obj )
     SpiInit(); //unique SPI
 }
 
-uint16_t SpiInOut(uint16_t outData )
+uint32_t SpiInOut(uint32_t outData )
 {   
     loraspi_mosi_write(outData);
     loraspi_control_start_write(1);
